adjust packages distributions
