+++ bb/bb_test.go
@@ -1,4 +1,4 @@
-package bank
+package bb
 
 import (
 	"testing"
