+++ bb/bb_validator.go
@@ -1,4 +1,4 @@
-package bank
+package bb
 
 import (
 	"strconv"
