+++ bb/request.go
@@ -1,4 +1,4 @@
-package letters
+package bb
 
 /*
 @author Philippe Moneda
