+++ caixa/caixa_test.go
@@ -1,4 +1,4 @@
-package bank
+package caixa
 
 import (
 	"testing"
